+++ src/components/Chart/index.ts
@@ -1,0 +1,1 @@
+export { default as OrigamChart } from './OrigamChart.vue'